+++ NEWS.md
@@ -2,11 +2,11 @@
 
 Changes in this minor release:
 
-* Update to sea object output to add event years.
+* Change to `sea` objects, added `event_years`.
 
-* Added function percent_scarred for calculating timeseries of number trees recording, scarred, and the percentage
+* Added function `percent_scarred()` for calculating timeseries of number trees recording, scarred, and the percentage.
 
-* Update to sample_depth to account for single-series fhx objects
+* Update to `sample_depth()` to account for single-series fhx objects
 
 # burnr v0.2.2